requirements.txt
aiofiles==23.1.0
aiogram==3.0.0rc1
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.0
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
certifi==2023.7.22
charset-normalizer==3.1.0
clients==1.4
emoji==2.6.0
exceptiongroup==1.1.1
frozenlist==1.3.3
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
magic-filter==1.0.9
multidict==6.0.4
pydantic==2.1.1
pydantic_core==2.4.0
python-dotenv==1.0.0
pytz==2023.3
sniffio==1.3.0
SQLAlchemy==2.0.16
typing_extensions==4.6.3
yarl==1.9.2
psycopg2-binary
